refactor: move LoggableError and implement Display trait
