+++ ampd/src/broadcaster.rs
@@ -23,10 +23,10 @@ use tracing::debug;
 use tracing::info;
 use valuable::Valuable;
 
-use crate::report::LoggableError;
 use crate::tofnd::grpc::SharableEcdsaClient;
 use crate::types::PublicKey;
 use dec_coin::DecCoin;
+use report::LoggableError;
 use tx::TxBuilder;
 
 pub mod accounts;
