+++ ampd/src/lib.rs
@@ -20,16 +20,16 @@ use tofnd::grpc::{MultisigClient, SharableEcdsaClient};
 use types::TMAddress;
 
 use crate::config::Config;
-use crate::report::Error;
+use crate::error::Error;
 
 mod broadcaster;
 pub mod config;
+pub mod error;
 mod event_processor;
 mod event_sub;
 mod evm;
 mod handlers;
 mod queue;
-pub mod report;
 pub mod state;
 mod tm_client;
 mod tofnd;
