+++ contracts/batching/Cargo.toml
@@ -8,4 +8,4 @@ edition = "2021"
 cosmwasm-schema = "1.1.3"
 cosmwasm-std = "1.1.3"
 cw2 = "0.15.1"
-thiserror = { version = "1.0.31" }
+thiserror = {workspace = true}
