+++ contracts/connection-router/Cargo.toml
@@ -38,7 +38,7 @@ flagset = { version = "0.4.3", features = ["serde"] }
 schemars = "0.8.10"
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 serde_json = "1.0.89"
-thiserror = { version = "1.0.31" }
+thiserror = {workspace = true}
 
 [dev-dependencies]
 cw-multi-test = "0.15.1"
