+++ contracts/multisig-prover/Cargo.toml
@@ -47,7 +47,7 @@ serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 serde_json = "1.0.89"
 service-registry = { workspace = true }
 sha3 = { version = "0.10", default-features = false, features = [] }
-thiserror = { version = "1.0.31" }
+thiserror = {workspace = true}
 
 [dev-dependencies]
 anyhow = "1.0"
