+++ contracts/multisig/Cargo.toml
@@ -42,7 +42,7 @@ getrandom = { version = "0.2", default-features = false, features = ["custom"] }
 schemars = "0.8.10"
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 serde_json = "1.0.89"
-thiserror = { version = "1.0.31" }
+thiserror = {workspace = true}
 
 [dev-dependencies]
 cw-multi-test = "0.15.1"
