+++ contracts/service-registry/Cargo.toml
@@ -36,7 +36,7 @@ cw-storage-plus = "0.15.1"
 cw2 = "0.15.1"
 schemars = "0.8.10"
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
-thiserror = { version = "1.0.31" }
+thiserror = {workspace = true}
 
 [dev-dependencies]
 cw-multi-test = "0.15.1"
