+++ contracts/voting-verifier/Cargo.toml
@@ -42,7 +42,7 @@ serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 serde_json = "1.0.89"
 service-registry = { workspace = true, features = ["library"] }
 sha3 = { version = "0.10", default-features = false, features = [] }
-thiserror = { version = "1.0.31" }
+thiserror = {workspace = true}
 
 [dev-dependencies]
 cw-multi-test = "0.15.1"
