+++ packages/axelar-wasm-std/Cargo.toml
@@ -38,7 +38,7 @@ schemars = "0.8.10"
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 serde_json = "1.0.89"
 sha3 = { version = "0.10", default-features = false, features = [] }
-thiserror = { version = "1.0.31" }
+thiserror = {workspace = true}
 
 [dev-dependencies]
 cw-multi-test = "0.15.1"
