+++ packages/report/Cargo.toml
@@ -7,3 +7,6 @@ edition = "2021"
 [dependencies]
 error-stack = { workspace = true }
 eyre = "0.6.8"
+valuable = { version = "0.1.0", features = ["derive"] }
+thiserror = { workspace = true }
+itertools = { workspace = true }
