+++ packages/report/src/lib.rs
@@ -1,3 +1,7 @@
+extern crate core;
+
+mod loggable;
 mod result_ext;
 
+pub use crate::loggable::LoggableError;
 pub use crate::result_ext::ResultCompatExt;
