+++ packages/report/src/loggable.rs
@@ -1,34 +1,13 @@
 use std::backtrace::Backtrace;
 use std::collections::VecDeque;
+use std::error::Error;
+use std::fmt::{Display, Formatter};
 use std::iter::FromIterator;
 use std::panic::Location;
 
 use error_stack::{AttachmentKind, Context, Frame, FrameKind, Report};
-use thiserror::Error;
-use tracing::error;
+use itertools::Itertools;
 use valuable::Valuable;
-
-#[derive(Error, Debug)]
-pub enum Error {
-    #[error("failed to load config, falling back on default")]
-    LoadConfig,
-    #[error("{0} is not a valid location to persist state")]
-    StateLocation(String),
-    #[error("event sub failed")]
-    EventSub,
-    #[error("event processor failed")]
-    EventProcessor,
-    #[error("broadcaster failed")]
-    Broadcaster,
-    #[error("state updater failed")]
-    StateUpdater,
-    #[error("tofnd failed")]
-    Tofnd,
-    #[error("connection failed")]
-    Connection,
-    #[error("task execution failed")]
-    Task,
-}
 
 #[derive(Valuable, PartialEq, Debug, Default)]
 pub struct LoggableError {
@@ -38,6 +17,29 @@ pub struct LoggableError {
     pub cause: Option<Box<LoggableError>>,
     pub backtrace: Option<LoggableBacktrace>,
 }
+
+impl Display for LoggableError {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        let msg = if self.attachments.is_empty() {
+            self.msg.clone()
+        } else {
+            let string = self
+                .attachments
+                .iter()
+                .map(|a| format!("{:?}", a))
+                .join("; ");
+            format!("{} ({:?})", self.msg, string)
+        };
+
+        let output = match &self.cause {
+            None => msg,
+            Some(cause) => format!("{}: {}", msg, cause),
+        };
+        write!(f, "{}", output)
+    }
+}
+
+impl Error for LoggableError {}
 
 #[derive(Valuable, PartialEq, Eq, Debug)]
 pub struct LoggableBacktrace {
@@ -143,5 +145,63 @@ impl<'a> From<&'a Frame> for FrameType<'a> {
             Attachment(Printable(p)) => FrameType::Printable(p.to_string()),
             Attachment(_) => unreachable!(),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::LoggableError;
+    use error_stack::Report;
+    use std::env;
+    use thiserror::Error;
+
+    #[derive(Error, Debug)]
+    enum Error {
+        #[error("{0}")]
+        FromString(String),
+    }
+
+    #[test]
+    fn correct_error_log() {
+        env::set_var("RUST_BACKTRACE", "1");
+        let line_offset = line!();
+        let report = Report::new(Error::FromString("error1".to_string()))
+            .attach_printable("foo1")
+            .change_context(Error::FromString("error2".to_string()))
+            .attach_printable("test1")
+            .attach_printable("test2")
+            .change_context(Error::FromString("error3".to_string()))
+            .attach(5);
+
+        let mut err = LoggableError::from(&report);
+
+        let root_err = err.cause.as_mut().unwrap().cause.as_mut().unwrap();
+
+        assert!(root_err.backtrace.is_some());
+        assert!(!root_err.backtrace.as_ref().unwrap().lines.is_empty());
+
+        root_err.backtrace = None;
+
+        let expected_err = LoggableError {
+            msg: "error3".to_string(),
+            attachments: vec!["opaque attachment".to_string()],
+            location: format!("packages/report/src/loggable.rs:{}:14", line_offset + 6),
+            cause: Some(Box::new(LoggableError {
+                msg: "error2".to_string(),
+                attachments: vec!["test1".to_string(), "test2".to_string()],
+                location: format!("packages/report/src/loggable.rs:{}:14", line_offset + 3),
+                cause: Some(Box::new(LoggableError {
+                    msg: "error1".to_string(),
+                    attachments: vec!["foo1".to_string()],
+                    location: format!("packages/report/src/loggable.rs:{}:22", line_offset + 1),
+                    cause: None,
+                    backtrace: None,
+                })),
+                backtrace: None,
+            })),
+            backtrace: None,
+        };
+
+        assert_eq!(err, expected_err);
     }
 }
